Update AKS installation script (#33)
* Remove nginx in favor of using Loadbalancers directly
* Remove deprecated/removed commands/flags
* Verify Coordinator and download root certificate as part of script
* Update CI

---------

+++ tools/aks_uninstall.sh
@@ -1,12 +1,10 @@
 #!/usr/bin/env bash
 
-okStatus="\e[92m\u221A\e[0m"
-warnStatus="\e[93m\u203C\e[0m"
-failStatus="\e[91m\u00D7\e[0m"
-
 helm uninstall -n emojivoto emojivoto
 kubectl delete ns emojivoto
-helm uninstall -n marblerun marblerun-coordinator
+helm uninstall -n marblerun marblerun
 kubectl delete ns marblerun
-helm uninstall -n kube-system nginx-ingress
-linkerd install --ignore-cluster | kubectl delete -f -
+if command -v linkerd &> /dev/null
+then
+    linkerd install --ignore-cluster | kubectl delete -f -
+fi
